BellmanFord + modif valeur de retour Visitor::Table()

+++ src/graph/algorithms/DFS.cpp
@@ -133,7 +133,7 @@ IGraph *DFS::G() const {
     return _G;
 }
 
-std::vector<int> DFS::table() {
+std::vector<double> DFS::table() {
     return _dfsnum;
 }
