+++ src/graph/algorithms/DetectedCycle.h
@@ -46,7 +46,7 @@ public:
 
     virtual IGraph *G() const override ;
 
-    virtual std::vector<int> table() override ;
+    virtual std::vector<double> table() override ;
 
 };
