+++ src/graph/algorithms/Kruskal.h
@@ -23,7 +23,7 @@ public:
     virtual void visit(DiGraph *g, Vertex *from);
     virtual void visit(FlowGraph *g, Vertex *from);
     virtual IGraph *G() const override;
-    virtual std::vector<int> table() override;
+    virtual std::vector<double> table() override;
 };
